Join client to room based on user ID
The client emits a handshake event containing the user ID.  The server adds the socket to the room keyed by user ID.  This is so that any messages addressed to the user can reach them via the room, avoiding other clients.  It also means a user can open multiple tabs and receive the same message.